+++ spec/super_settings/setting_spec.rb
@@ -577,6 +577,17 @@ describe SuperSettings::Setting do
           expect(new_setting.value).to eq "foobar"
           expect(new_setting.deleted).to be(false)
         end
+
+        it "should silently skip deleting a key that does not exist" do
+          success, settings = SuperSettings::Setting.bulk_update([
+            {
+              key: "nonexistent",
+              deleted: true
+            }
+          ])
+          expect(success).to eq true
+          expect(settings).to be_empty
+        end
       end
 
       describe "set" do